+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## `1.0.65-SNAPSHOT`
+## `1.0.70-SNAPSHOT`
 
 Release date `UNRELEASED`
+
+## `1.0.69-rc1`
+
+Release date `2023-01-13`
 
 - Implement `types/bean*` for conversion of Java classes to maps.
 - Define handler types as symbols.